# api/openapi.yaml
openapi: 3.0.3
info:
  title: Helling API
  version: 0.2.0
  description: |
    Helling is a self-hosted, single-node container and VM management platform.
    This is the v0.2 Platform Core bridge: authentication, instance CRUD,
    operation polling, schedules, webhooks, host firewall, events, audit, and
    user Incus scope assignment.

    All endpoints under `/v1/` are stable within v0.2. Breaking changes ship under `/v2/`.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Bizarre Industries
    url: https://github.com/Bizarre-Industries/helling
servers:
  - url: https://localhost/v1
    description: Local installation behind the Caddy edge service
  - url: http+unix:///run/helling/api.sock/v1
    description: Direct socket access (CLI, dev)
...
security:
  - cookieAuth: []
paths:
...
  /auth/login:
    post:
      operationId: authLogin
      tags:
        - auth
      security: []
      summary: Exchange username + password for a session
      description: |
        Validates credentials and, on success, issues a session via the
        `helling_session` cookie. CLI and web clients use the JSON envelope's
        Ed25519-signed access token. Accounts with enabled TOTP receive a
        pre-session MFA challenge instead of a session.
      requestBody:
        required: true
        description: User credentials
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful. Sets `helling_session` cookie and returns an access token for API clients.
          headers:
            Set-Cookie:
              description: Session cookie set on successful login.
              schema:
                type: string
                description: Standard Set-Cookie header value.
                example: helling_session=...; Path=/; HttpOnly; Secure; SameSite=Lax; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenEnvelope"
              example:
                data:
                  access_token: eyJhbGciOiJFZERTQSJ9...
                  token_type: Bearer
                  expires_in: 900
                  mfa_required: false
        "202":
          description: Password accepted, but MFA must be completed before a session is issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenEnvelope"
              example:
                data:
                  mfa_required: true
                  mfa_token: c08f4a7281b54d9abef67c210d3e49a2
        "204":
          description: Browser-compatible login successful response for cookie-only clients.
          headers:
            Set-Cookie:
              description: Session cookie set on successful login.
              schema:
                type: string
                description: Standard Set-Cookie header value.
                example: helling_session=...; Path=/; HttpOnly; Secure; SameSite=Lax; Max-Age=604800
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/setup/status:
    get:
      operationId: authSetupStatus
      tags:
        - auth
      security: []
      summary: Report whether first-admin setup is required
      description: Returns whether the daemon currently has zero users and will accept one-time setup.
      responses:
        "200":
          description: Setup status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupStatus"
              example:
                setup_required: true
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/setup:
    post:
      operationId: authSetup
      tags:
        - auth
      security: []
      summary: Create the first admin user
      description: |
        One-time first-boot setup. This endpoint is available only while no
        users exist and requires the installer setup token from
        `/etc/helling/setup-token`.
      requestBody:
        required: true
        description: First admin credentials and installer setup token
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "201":
          description: First admin user created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: 1
                username: admin
                is_admin: true
                created_at: "2026-05-04T12:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/mfa/complete:
    post:
      operationId: authMfaComplete
      tags:
        - auth
      security: []
      summary: Complete a pre-session MFA challenge
      description: |
        Validates a TOTP code or one-time recovery code against the pre-session
        challenge returned by `/auth/login`. A successful request issues the
        same session cookie and access-token envelope as password-only login.
      requestBody:
        required: true
        description: MFA challenge token and TOTP or recovery code.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFACompleteRequest"
            example:
              mfa_token: c08f4a7281b54d9abef67c210d3e49a2
              totp_code: "123456"
      responses:
        "200":
          description: MFA successful. Sets `helling_session` cookie and returns an access token for API clients.
          headers:
            Set-Cookie:
              description: Session cookie set on successful MFA completion.
              schema:
                type: string
                description: Standard Set-Cookie header value.
                example: helling_session=...; Path=/; HttpOnly; Secure; SameSite=Lax; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenEnvelope"
              example:
                data:
                  access_token: eyJhbGciOiJFZERTQSJ9...
                  token_type: Bearer
                  expires_in: 900
                  mfa_required: false
        "204":
          description: MFA successful cookie-only response.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/logout:
    post:
      operationId: logout
      tags:
        - auth
      summary: Invalidate the current session
      description: Revokes the session bound to the `helling_session` cookie and clears it client-side.
      responses:
        "204":
          description: Session invalidated. Clears `helling_session` cookie.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags:
        - auth
      summary: Identity of the authenticated user
      description: Returns the profile of the user bound to the current session cookie.
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /instances:
    get:
      operationId: listInstances
      tags:
        - instances
      summary: List instances visible to the user
      description: Returns the instances visible to the authenticated user, optionally filtered by status.
      parameters:
        - in: query
          name: status
          description: Optional status filter
          schema:
            type: string
            description: Lifecycle state to filter by.
            enum: [running, stopped, frozen, error]
            example: running
      responses:
        "200":
          description: List of instances
          content:
            application/json:
              schema:
                type: array
                description: Array of instance records.
                items:
                  $ref: "#/components/schemas/Instance"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createInstance
      tags:
        - instances
      summary: Create a new instance (async)
      description: Submits an instance create request. Returns the tracking operation; poll `/operations/{id}` for completion.
      requestBody:
        required: true
        description: Instance create parameters
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstanceCreate"
      responses:
        "202":
          description: Operation accepted; poll `/operations/{id}` for completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /instances/{name}:
    parameters:
      - $ref: "#/components/parameters/InstanceName"
    get:
      operationId: getInstance
      tags:
        - instances
      summary: Fetch a single instance
      description: Returns the current snapshot of one instance.
      responses:
        "200":
          description: Instance details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteInstance
      tags:
        - instances
      summary: Delete an instance (async)
      description: Submits an instance delete request. Returns the tracking operation; poll `/operations/{id}` for completion.
      responses:
        "202":
          description: Deletion operation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /instances/{name}/start:
    parameters:
      - $ref: "#/components/parameters/InstanceName"
    post:
      operationId: startInstance
      tags:
        - instances
      summary: Start an instance (async)
      description: Submits a start request for the named instance. Returns the tracking operation.
      responses:
        "202":
          description: Start operation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /instances/{name}/stop:
    parameters:
      - $ref: "#/components/parameters/InstanceName"
    post:
      operationId: stopInstance
      tags:
        - instances
      summary: Stop an instance (async)
      description: Submits a stop request for the named instance. Optional body controls force + timeout. Returns the tracking operation.
      requestBody:
        required: false
        description: Optional stop options (force flag, timeout seconds)
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstanceStopRequest"
      responses:
        "202":
          description: Stop operation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /operations:
    get:
      operationId: listOperations
      tags:
        - operations
      summary: List operations for the current user
      description: Returns the recent operations submitted by the authenticated user, optionally filtered by status.
      parameters:
        - in: query
          name: status
          description: Optional status filter for operations.
          schema:
            type: string
            description: Operation status to filter by.
            enum: [pending, running, success, failure, cancelled]
            example: running
        - in: query
          name: limit
          description: Maximum number of results to return.
          schema:
            type: integer
            description: Maximum result count.
            minimum: 1
            maximum: 200
            default: 50
            example: 50
      responses:
        "200":
          description: Recent operations
          content:
            application/json:
              schema:
                type: array
                description: Array of operation records.
                items:
                  $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /operations/{id}:
    parameters:
      - in: path
        name: id
        required: true
        description: Operation identifier (UUID).
        schema:
          type: string
          description: UUID v7 string.
          format: uuid
          example: 018f3a40-0000-7000-8000-000000000000
    get:
      operationId: getOperation
      tags:
        - operations
      summary: Fetch a single operation by ID
      description: Returns the current state of a single operation by its UUID.
      responses:
        "200":
          description: Operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /users:
    get:
      operationId: listUsers
      tags:
        - users
      summary: List local users
      description: Returns all local Helling users. Admin role required.
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListEnvelope"
              example:
                data:
                  - id: 1
                    username: admin
                    is_admin: true
                    incus_project: default
                    created_at: "2026-05-03T12:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createUser
      tags:
        - users
      summary: Create a local user
      description: Creates a Helling user and provisions its delegated Incus scope. Admin role required.
      requestBody:
        required: true
        description: User creation request.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
            example:
              username: alice
              password: correct horse battery staple
              is_admin: false
              incus_project: alice
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
              example:
                data:
                  id: 2
                  username: alice
                  is_admin: false
                  incus_project: alice
                  created_at: "2026-05-05T09:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /users/{id}/scope:
    parameters:
      - $ref: "#/components/parameters/UserID"
    put:
      operationId: updateUserScope
      tags:
        - users
      summary: Assign a user to an Incus project
      description: Replaces the user's Incus project scope and rotates the restricted trust entry. Admin role required.
      requestBody:
        required: true
        description: Scope assignment request.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserScopeRequest"
            example:
              scope: incus:project:alice
      responses:
        "200":
          description: Scope assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserScopeEnvelope"
              example:
                data:
                  scope: incus:project:alice
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /firewall/host/{id}:
    parameters:
      - $ref: "#/components/parameters/UUID"
    delete:
      operationId: deleteHostFirewallRule
      tags:
        - firewall
      summary: Delete a host firewall rule
      description: Deletes a Helling-managed nftables rule by metadata id. Admin role required.
      responses:
        "204":
          description: Host firewall rule deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: helling_session
      description: Session cookie issued by `POST /auth/login`. Sent automatically by the browser; CLI stores it under `~/.config/helling/session`.
...
  schemas:
...
    LoginRequest:
      type: object
      description: Username + password credentials submitted to `/auth/login`.
      required: [username, password]
      properties:
        username:
          type: string
          description: Account username.
          minLength: 1
          maxLength: 64
          example: alice
        password:
          type: string
          description: Account password (cleartext over TLS only).
          minLength: 8
          maxLength: 256
          format: password
          example: correct horse battery staple
...
    AuthTokenEnvelope:
      type: object
      description: JSON envelope returned by auth flows that issue or continue a session.
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuthTokenResponse"
...